Added the "Would you like a drink?" project
Still a work in progress. I thought it would be a good idea to try out an API to develop a website of some sort, and so I did it.
This uses TheCockTailDB's API to fetch drinks however you like. It provides the preparation, ingredients and even images of many beverages, including non-alcoholic ones.